weather/requirements.txt
Django>=4.2.0
python-dotenv>=1.0.0
requests>=2.31.0
gunicorn>=21.2.0
whitenoise>=6.5.0
django-crispy-forms>=2.1
crispy-bootstrap5>=0.7
